Versão sem controller para o login

+++ lib/model/carteira.dart
@@ -1,5 +1,5 @@
 class Carteira {
-  double saldo;
+  int saldo;
   String id;
 
   Carteira({
